Various tweaks and fixes to distribution examples
