+++ pints/toy/__init__.py
@@ -11,10 +11,9 @@ from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 from ._logistic import LogisticModel            # noqa
 from ._hh_ik import HodgkinHuxleyIKModel        # noqa
-from ._logpdfs import (
+from ._distributions import ( # noqa
     BimodalMultivariateGaussianLogPDF,
     MultivariateGaussianLogPDF,
     RosenbrockLogPDF,
     TwistedGaussianLogPDF,
-) # noqa
-
+)
